beginning of spawner spawner

+++ Assets/Scripts/GameParameters.cs
@@ -8,7 +8,7 @@ public class GameParameters : MonoBehaviour {
     private long playerScore;
     private const long maxScore = 999999999999L;
     public Vector3 lastCheckpoint;
-    public int speedMultiplier = 1;
+    public float speedMultiplier = 1;
 
     void Awake()
     {
